Continuing persistence

+++ backend/include/myroomies/utils/db/Def.h
@@ -6,7 +6,7 @@ namespace myroomies {
 namespace utils {
 namespace db {
 
-typedef uint64_t Key_t;
+typedef int Key_t;
 typedef Key_t ForeignKey_t;
 
 } /* namespace db */
